pyproject.toml
[tool.poetry]
name = "repl_python3_beta_obs-all"
version = "0.1.0"
description = ""
authors = ["StanleyBlack <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
